remove parser core rollout scaffolding
